Factor out `SCMBasedRetriever` roughly backporting https://github.com/jenkinsci/pipeline-groovy-lib-plugin/pull/57/files/840cd6004501d8d9e52317288ebbbb86fddbd7a4..444bde7967e0fa5ad9242bb1c2e08a073b49f21e + https://github.com/jenkinsci/pipeline-groovy-lib-plugin/pull/57/commits/e5e95d0d2f92f5091f1646915f82a88e7e2a5590
