+++ src/main/resources/org/jenkinsci/plugins/workflow/libs/SCMBasedRetriever/options.properties
@@ -20,6 +20,4 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 
-blurb=\
-    Loads a library from an SCM plugin using newer interfaces optimized for this purpose. \
-    The recommended option when available.
+libraryPath=Library Path (optional)
