+++ src/main/resources/org/jenkinsci/plugins/workflow/libs/SCMRetriever/config.properties
@@ -20,7 +20,6 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 
-libraryPath=Library Path (optional)
 blurb=\
     Loads a library from any SCM plugin supported by Jenkins Pipeline, for example in the <code>checkout</code> step. \
     Not all version expressions may be supported, no prevalidation of versions is available, \
