+++ src/main/resources/org/jenkinsci/plugins/workflow/libs/SCMSourceRetriever/config.jelly
@@ -24,13 +24,8 @@ THE SOFTWARE.
 -->
 
 <?jelly escape-by-default='true'?>
-<j:jelly xmlns:j="jelly:core" xmlns:f="/lib/form">
+<j:jelly xmlns:j="jelly:core" xmlns:f="/lib/form" xmlns:st="jelly:stapler">
     <f:description>${%blurb}</f:description>
     <f:dropdownDescriptorSelector field="scm" descriptors="${descriptor.SCMDescriptors}" title="${%Source Code Management}"/>
-    <f:entry field="clone" title="${%Fresh clone per build}">
-        <f:checkbox/>
-    </f:entry>
-    <f:entry field="libraryPath" title="${%libraryPath}">
-        <f:textbox checkMethod="post"/>
-    </f:entry>
+    <st:include page="options" class="${descriptor.clazz}" from="${instance}"/>
 </j:jelly>
