--- src/main/resources/org/jenkinsci/plugins/workflow/libs/SCMSourceRetriever/help-libraryPath.html
@@ -1,5 +1,0 @@
-<div>
-    A relative path from the root of the SCM to the root of the library.
-    Leave this field blank if the root of the library is the root of the SCM.
-    Note that ".." is not permitted as a path component to avoid security issues.
-</div>
